+++ src/main/java/org/goormthon/seasonthon/nocheongmaru/domain/feed/service/FeedService.java
@@ -18,13 +18,12 @@ import org.goormthon.seasonthon.nocheongmaru.domain.member.entity.Member;
 import org.goormthon.seasonthon.nocheongmaru.domain.member.repository.MemberRepository;
 import org.goormthon.seasonthon.nocheongmaru.domain.mission.entity.MemberMission;
 import org.goormthon.seasonthon.nocheongmaru.domain.mission.entity.Mission;
-import org.goormthon.seasonthon.nocheongmaru.domain.mission.repository.MemberMissionRepository;
-import org.goormthon.seasonthon.nocheongmaru.domain.mission.repository.MissionRepository;
+import org.goormthon.seasonthon.nocheongmaru.domain.mission.repository.membermission.MemberMissionRepository;
+import org.goormthon.seasonthon.nocheongmaru.domain.mission.repository.mission.MissionRepository;
 import org.goormthon.seasonthon.nocheongmaru.global.exception.member.AlreadyUploadedTodayException;
 import org.goormthon.seasonthon.nocheongmaru.global.exception.member.FeedNotFoundException;
 import org.goormthon.seasonthon.nocheongmaru.global.exception.member.ForbiddenFeedAccessException;
 import org.goormthon.seasonthon.nocheongmaru.global.exception.member.MemberNotFoundException;
-import org.goormthon.seasonthon.nocheongmaru.global.exception.member.MissionNotFoundException;
 import org.goormthon.seasonthon.nocheongmaru.global.exception.member.TodayMissionAccessDeniedException;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
