+++ src/main/java/org/goormthon/seasonthon/nocheongmaru/domain/mission/entity/Status.java
@@ -1,0 +1,8 @@
+package org.goormthon.seasonthon.nocheongmaru.domain.mission.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum Status {
+    ASSIGNED, COMPLETED
+}
