+++ src/main/java/org/goormthon/seasonthon/nocheongmaru/domain/notification/scheduler/MissionAssignmentScheduler.java
@@ -8,8 +8,8 @@ import org.goormthon.seasonthon.nocheongmaru.domain.member.entity.Member;
 import org.goormthon.seasonthon.nocheongmaru.domain.member.repository.MemberRepository;
 import org.goormthon.seasonthon.nocheongmaru.domain.mission.entity.MemberMission;
 import org.goormthon.seasonthon.nocheongmaru.domain.mission.entity.Mission;
-import org.goormthon.seasonthon.nocheongmaru.domain.mission.repository.MemberMissionRepository;
-import org.goormthon.seasonthon.nocheongmaru.domain.mission.repository.MissionRepository;
+import org.goormthon.seasonthon.nocheongmaru.domain.mission.repository.membermission.MemberMissionRepository;
+import org.goormthon.seasonthon.nocheongmaru.domain.mission.repository.mission.MissionRepository;
 import org.goormthon.seasonthon.nocheongmaru.domain.notification.entity.NotificationType;
 import org.goormthon.seasonthon.nocheongmaru.domain.notification.service.NotificationService;
 import org.springframework.scheduling.annotation.Scheduled;
